cleaner/requirements.txt
gitpython
colorama
pathlib
requests
